Have the palette code be in its own file.
