+++ src/utils/Makefile.am
@@ -26,6 +26,8 @@ COMMON_SOURCES += filters_utils.c
 COMMON_SOURCES += filters_utils.h
 COMMON_SOURCES += huffman_utils.c
 COMMON_SOURCES += huffman_utils.h
+COMMON_SOURCES += palette.c
+COMMON_SOURCES += palette.h
 COMMON_SOURCES += quant_levels_dec_utils.c
 COMMON_SOURCES += quant_levels_dec_utils.h
 COMMON_SOURCES += rescaler_utils.c
